requirements.txt
python-telegram-bot==20.0
google-generativeai==0.8.3
python-dotenv==1.0.0
langchain_google_genai
langchain
pyjwt
requests
